backend/requirements.txt
email-validator==2.3.0
fastapi==0.120.4
psycopg2-binary==2.9.11
pydantic==2.12.3
uvicorn==0.38.0
pandas==2.2.3
openpyxl==3.1.5
sentence-transformers
python-dateutil
python-multipart
bottleneck>=1.3.6
